+++ config.php
@@ -1,0 +1,10 @@
+<?php
+return [
+    'db' => [
+        'host' => 'localhost',
+        'user' => 'root',
+        'password' => '',
+        'database' => 'yeticave',
+    ],
+    'mailer' => []
+];
