--- data.php
@@ -1,52 +1,0 @@
-<?php
-
-$isAuth = rand(0, 1);
-
-$userName = 'Наталья';
-
-$categories = ["Доски и лыжи", "Крепления", "Ботинки", "Одежда", "Инструменты", "Разное"];
-
-$lots = [
-    [
-        "title" => "2014 Rossignol District Snowboard",
-        "category" => $categories[0],
-        "price" => 10999,
-        "img_url" => "img/lot-1.jpg",
-        "finish_date" => "2024-12-12"
-    ],
-    [
-        "title" => "DC Ply Mens 2016/2017 Snowboard",
-        "category" => $categories[0],
-        "price" => 159999,
-        "img_url" => "img/lot-2.jpg",
-        "finish_date" => "2024-12-14"
-    ],
-    [
-        "title" => "Крепления Union Contact Pro 2015 года размер L/XL",
-        "category" => $categories[1],
-        "price" => 8000,
-        "img_url" => "img/lot-3.jpg",
-        "finish_date" => "2024-12-15"
-    ],
-    [
-        "title" => "Ботинки для сноуборда DC Mutiny Charocal",
-        "category" => $categories[2],
-        "price" => 10999,
-        "img_url" => "img/lot-4.jpg",
-        "finish_date" => "2024-12-16"
-    ],
-    [
-        "title" => "Куртка для сноуборда DC Mutiny Charocal",
-        "category" => $categories[3],
-        "price" => 7500,
-        "img_url" => "img/lot-5.jpg",
-        "finish_date" => "2024-12-17"
-    ],
-    [
-        "title" => "Маска Oakley Canopy",
-        "category" => $categories[5],
-        "price" => 5400,
-        "img_url" => "img/lot-6.jpg",
-        "finish_date" => "2024-12-18"
-    ],
-];
