+++ functions/template.php
@@ -1,6 +1,5 @@
 <?php
 
-
 /**
  * Подсчитывает время до окончания показа лота
  * @param string $date
@@ -36,3 +35,26 @@ function formatPrice(int|float $price): string
     return $price . ' ₽';
 }
 
+/**
+ * Подключает шаблон, передает туда данные и возвращает итоговый HTML контент
+ * @param string $name Путь к файлу шаблона относительно папки templates
+ * @param array $data Ассоциативный массив с данными для шаблона
+ * @return string Итоговый HTML
+ */
+function includeTemplate(string $name, array $data = []): string
+{
+    $name = 'templates/' . $name;
+    $result = '';
+
+    if (!is_readable($name)) {
+        return $result;
+    }
+
+    ob_start();
+    extract($data);
+    require $name;
+
+    $result = ob_get_clean();
+
+    return $result;
+}
