+++ init.php
@@ -1,0 +1,5 @@
+<?php
+
+require_once 'functions/template.php';
+require_once 'functions/validators.php';
+require_once 'functions/db.php';
